(Internal function) Perform the clustering process of IPCAPS

(Internal function) Perform the clustering process of IPCAPS

clustering(
  dataframe,
  node = 1,
  result.dir,
  threshold,
  min.fst,
  method = "mix",
  min.in.group = 20,
  datatype = "snp",
  nonlinear = FALSE
)

Arguments

dataframe
A data frame containing raw.data (matrix or data frame), label (vector), and index (vector). raw.data represents a matrix of subset of input data. label represents a vector of labels for all rows of raw.data. index represents a vector of indexes that are selected from the original input data.

node
An integer representing the current node number which is being analyzed.

result.dir
An output directory

threshold
A threshold or a cutoff to stop the IPCAPS process. Also see ipcaps (the parameter threshold).

min.fst
A number represents a cutoff for mininum Fst value.

method
A clustering method selected from the ipcaps function. See ipcaps for available methods.

min.in.group
A integer represents a minimum number of group members.

datatype
To specify whether the input data are 'snp' or other type. Defalut = 'snp'.

nonlinear
(Unimplemented) To specify whether linear or non-linear method is used for IPCAPS analysis. If TRUE, non-linear method is used, otherwise linear method is used. Default = FALSE.

Value

A list containing status, node, and new.index as explained below:

$status is either 0 representing that the criterion is not met, or 1 representing that the criterion is met.

$node is an integer representing the current node number which is being analyzed.

$new.index is a list of vectors containing a subset of indexes split from dataframe$index according to a clustering result.